requirements.txt
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-openai==0.2.6
langgraph==0.2.44
rich==13.9.4
python-dotenv==1.0.1
ddgs==6.3.5
openai==1.54.3